+++ Jumpscale/core/BASECLASSES/test/base_tests.py
@@ -1,5 +1,4 @@
 import unittest
-import subprocess
 from loguru import logger
 from uuid import uuid4
